update demo project on ui role.

+++ demo/public/js/views/ProjectContact.js
@@ -12,7 +12,7 @@ define(['text!templates/projectContact.html'],function(contactTemplate){
 		},
 
 		initialize: function(options){
-			this.pid = options.pid;
+			this.project = options.project;
 			if(options.addButton){
 				this.addButton = options.addButton;
 			}
@@ -24,7 +24,7 @@ define(['text!templates/projectContact.html'],function(contactTemplate){
 		addContact: function(){
 			var $responseArea = this.$('.actionArea');
 
-			$.post('/projects/'+ this.pid +'/contacts'
+			$.post('/projects/'+ this.project._id +'/contacts'
 				,{contactId: this.model.get('_id')}
 				,function onSuccess(){
 					$responseArea.text('已添加成功！');
@@ -36,25 +36,28 @@ define(['text!templates/projectContact.html'],function(contactTemplate){
 		},
 
 		removeContact: function(){
-			var $responseArea = this.$('.actionArea');
-			$responseArea.text('正在移除....');
-			$.ajax({
-					url: '/projects/'+ this.pid +'/contacts/' + this.model.get('_id')
-					,type: 'DELETE'
-					,data: {
-						contactId: this.model.get('accountId')
-					}
-				}).done(function onSuccess(){
-					$responseArea.text('移除成功！')
-				}).fail(function onError(){
-					$responseArea.text('移除失败');
-				});
-			return false;
+			if(confirm('确认移除'+ this.model.get('username')+'用户吗？')){
+				var $responseArea = this.$('.actionArea');
+				$responseArea.text('正在移除....');
+				$.ajax({
+						url: '/projects/'+ this.project._id +'/contacts/' + this.model.get('_id')
+						,type: 'DELETE'
+						,data: {
+							contactId: this.model.get('accountId')
+						}
+					}).done(function onSuccess(){
+						$responseArea.text('移除成功！')
+					}).fail(function onError(){
+						$responseArea.text('移除失败');
+					});
+				return false;
+			}
 		},
 
 
 		render: function(){
 			this.$el.html(this.template({
+				project: this.project.toJSON(),
 				model: this.model.toJSON(),
 				addButton: this.addButton,
 				removeButton: this.removeButton
